Fix spec for lts-12.2
This fix reveals the suspect use of `Vector 0`'s `Monoid` instance in
this library. `Vector 0` is a fine monoid to utilize for an empty list
of emails, but `Vector (Suc n)` is problematic. However `EmailAddress`
has rightfully dropped its `Monoid` instance and as such this usage now
fails as `Vector` relies on the `Monoid` instance of its contents as its
`mempty` is `replicate nat mempty`.